Create simpler Receiver code
The JeeNode_Receive code was mostly based
on the rf12 demo from JeeLabs, which included
a lot of code we don't need. I created a
version with all of that stripped out.

Also copied the RF12 config info from the
Send sketch to the new Receive sketch.

The Receiver is no longer configured via EEPROM,
but the Sender was using hardcoded config anyways,
so there wasn't much need for EEPROM config.